fix(date-basis): second review round — layering, boundary validation, persona test
- Layering: DATE_BASIS_* constants and validate_date_basis now live in
  shared/feedback.py (the data layer); shared/api.py re-exports them so
  every existing import keeps working. shared.feedback no longer pulls
  API-resolver machinery into job Lambdas, and the potential
  api<->feedback import cycle is gone.
- research_step_handler validates date_basis on consume: Step Functions
  execution input is an unvalidated boundary (anyone with
  StartExecution can pass arbitrary JSON), so the comment-level trust
  in upstream validation is replaced with validate_date_basis.
- Persona generation consumption pinned by a test: the job Lambda
  forwards the filters dict verbatim into generate_personas (whose
  get_feedback_context unpacks date_basis) — the reviewer couldn't see
  this in the truncated diff because no persona-generator change was
  needed; the new test guards against a future rebuild of the dict
  dropping the field.
